refactor v1 structure + add financial page
