+++ courses/serializers/v1/departments_test.py
@@ -1,0 +1,4 @@
+import pytest
+
+
+pytestmark = [pytest.mark.django_db]
